requirements.txt
setuptools
beautifulsoup4
bs4
celery
Django
django-allauth
html5lib
lxml
PyJWT
redis
requests
six
url-normalize
urllib3 
w3lib
requests-html
urllib3
whitenoise
psycopg2-binary
djangorestframework